use setTextAndSelection instead of maybeSetText
maybeSetText is used to keep in sync updates between c++ fabric state
and Java.
setTextAndSelection is used to over-ride from JavaScript the Native
value (saved in memory in iOS objc and Android Java) of the text or the
cursor selection.

to update the error onChangeText callback we need to trigger this
over-ride (setNativeProps) of the errorMessage

the logic is similar to maybeSetText, but no changes to cpp APi are
required. The solution is the same for iOS and Android.